Next

+++ ow.Framework/Game/Enums/GateStatus.cs
@@ -1,6 +1,6 @@
 ﻿namespace ow.Framework.Game.Enums
 {
-    public enum GateStatus : byte
+    public enum GateStatus : uint
     {
         Offline = 0,
         Online = 1
